packages/backend/services/image-tagger/requirements.txt
# ===== Core ML =====
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.24.0
transformers>=4.37.0
scikit-learn>=1.3.0
piexif>=1.1.3
tqdm>=4.66.0

# ===== Web Stack (FastAPI) =====
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart>=0.0.6

# ===== Utilities =====
python-dotenv>=1.0.0
requests>=2.31.0